Moved MPIArray into mpiarray.py in preparation for removal of mpidataset.
